[MWX Checkpoint] Before editing src/io/pytorch_loader.rs (2026-06-13T09-24-00-288Z)
Former-commit-id: 6ad0631a3eda615682cd152a65f30a2075e9f3a8

+++ src/io/pytorch_loader.rs
@@ -189,8 +189,8 @@ enum TensorReadResult<'a> {
 
 fn try_read_tensor<'a>(data: &mut &'a [u8]) -> Result<TensorReadResult<'a>, ()> {
     let tag = read_varint(data).ok_or(())?;
-    let outer_wire = tag & 0x07;
-    if outer_wire != WireType::LengthDelimited as u64 { return Err(()); }
+    let outer_wire = (tag & 0x07) as u8;
+    if outer_wire != WireType::LengthDelimited as u8 { return Err(()); }
     let payload_len = read_varint(data).ok_or(())? as usize;
     if data.len() < payload_len { return Err(()); }
     let payload = &data[..payload_len];
